pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bq-auditor"
version = "0.1.0"
description = "BigQuery cost optimization and governance audit CLI tool"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
authors = [{ name = "BQ-Auditor Contributors" }]
keywords = ["bigquery", "cost", "audit", "data-engineering", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Database",
    "Topic :: Utilities",
]
dependencies = [
    "typer>=0.12.0",
    "pydantic>=2.0.0",
    "google-cloud-bigquery>=3.0.0",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pytest>=8.0.0",
    "pytest-mock>=3.14.0",
    "types-PyYAML>=6.0.0",
]

[project.scripts]
bq-auditor = "bq_auditor.main:app"

[project.urls]
Homepage = "https://github.com/your-org/bq-auditor"
Repository = "https://github.com/your-org/bq-auditor"
Issues = "https://github.com/your-org/bq-auditor/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/bq_auditor"]

[tool.ruff]
target-version = "py39"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "ANN", # flake8-annotations
    "RUF", # ruff-specific
]
ignore = [
    "ANN101", # Missing type annotation for `self`
    "ANN102", # Missing type annotation for `cls`
    "RUF001", # String contains ambiguous unicode (日本語 UI 文字列で許容)
    "RUF002", # Docstring contains ambiguous unicode (日本語 docstring で許容)
    "RUF003", # Comment contains ambiguous unicode (日本語コメントで許容)
]

[tool.ruff.lint.isort]
known-first-party = ["bq_auditor"]

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
mypy_path = "src"

[[tool.mypy.overrides]]
module = "google.cloud.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
